requirements.txt
streamlit>=1.44.0
pdfplumber>=0.11.6
pandas>=2.2.0
python-dotenv>=1.0.1
google-generativeai>=0.5.0
python-dotenv>=1.0.1
openai>=1.10.0
